+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/AppConfig.kt
@@ -31,6 +31,7 @@ import io.sentry.android.gradle.util.hookWithPackageTasks
 import io.sentry.android.gradle.util.info
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.file.Directory
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.internal.build.event.BuildEventListenerRegistryInternal
@@ -68,11 +69,32 @@ fun AppExtension.configure(
             buildEvents
         )
 
+        // TODO: do this only once, and all other tasks should be SentryVariant.configureSomething
+        val sentryVariant = if (isAGP74) null else AndroidVariant70(variant)
+        sentryVariant?.configureNativeSymbolsTask(
+            project,
+            extension,
+            sentryTelemetryProvider,
+            cliExecutable,
+            sentryOrg,
+            sentryProject
+        )
+
+        val additionalSourcesProvider = project.provider {
+            extension.additionalSourceDirsForSourceContext.getOrElse(emptySet())
+                .map { project.layout.projectDirectory.dir(it) }
+        }
+        val sourceFiles = sentryVariant?.sources(
+            project,
+            additionalSourcesProvider
+        )
+
         val tasksGeneratingProperties = mutableListOf<TaskProvider<out PropertiesFileOutputTask>>()
         val sourceContextTasks = variant.configureSourceBundleTasks(
             project,
             extension,
             sentryTelemetryProvider,
+            sourceFiles,
             cliExecutable,
             sentryOrg,
             sentryProject
@@ -96,17 +118,6 @@ fun AppExtension.configure(
             sentryProject
         )
         generateProguardUuidTask?.let { tasksGeneratingProperties.add(it) }
-
-        // TODO: do this only once, and all other tasks should be SentryVariant.configureSomething
-        val sentryVariant = if (isAGP74) null else AndroidVariant70(variant)
-        sentryVariant?.configureNativeSymbolsTask(
-            project,
-            extension,
-            sentryTelemetryProvider,
-            cliExecutable,
-            sentryOrg,
-            sentryProject
-        )
 
         variant.configureDebugMetaPropertiesTask(
             project,
@@ -187,6 +198,7 @@ private fun ApplicationVariant.configureSourceBundleTasks(
     project: Project,
     extension: SentryPluginExtension,
     sentryTelemetryProvider: Provider<SentryTelemetryService>,
+    sourceFiles: Provider<out Collection<Directory>>?,
     cliExecutable: Provider<String>,
     sentryOrg: String?,
     sentryProject: String?
@@ -208,6 +220,7 @@ private fun ApplicationVariant.configureSourceBundleTasks(
             sentryTelemetryProvider,
             variant,
             paths,
+            sourceFiles,
             cliExecutable,
             sentryOrg,
             sentryProject,
@@ -289,6 +302,11 @@ private fun ApplicationVariant.configureProguardMappingsTasks(
                     extension,
                     sentryTelemetryProvider,
                     output = outputDir,
+                    proguardMappingFile = SentryTasksProvider.getMappingFileProvider(
+                        project,
+                        variant,
+                        dexguardEnabled
+                    ),
                     taskSuffix = name.capitalized
                 )